requirements.txt
pyspark
delta-spark
apache-airflow
pandas
streamlit
deltalake